chore(gcs-conformance): refactor stream so retryListener is internal
